Move the definitions of distance metrics/measures to DistanceMeasure.scala or DistanceMetric.scala
